Fix "out of tree" builds - src/ceph.in ceph_ver.h .git_version, and PEP 440 fix.
When building out of tree, ceph.in is really $(srcdir)/ceph.in .
If we're building with gnu make, that can be shortened to just $<.

When building out of tree:
 - set GIT_DIR to point to source tree .git.
 - reference .git_version as $(src)/.git_version
 - never call ceph_ver.h ./ceph_ver.h

.git_version is part of "dist", so for consistency it should be in the
	source tree.
ceph_ver.h is a generated file; it belongs in the build tree.
	./ceph_ver.h is not the same thing as ceph_ver.h to make,
	and it's silly to prefix targets with ./

It would be better if configure made .git_version directly and the
constructing logic just weren't in the build tree.  But I didn't do that.

Instead of Looking for ceph_ver.h, this looks for .git_version .
Same information basically.  Ideally this should know how to look in
the build tree for ceph_ver.h, but that was more python mongery than I
wanted to hack.

Once this gets the result of "git describe", I've added logic to try
to format the version string to be PEP 440 friendly.  So, instead of
10.0.3-1027-ge51fd74, this will make 10.0.3.post10+ge51fd74 .  This fixes
a warning message at build time for me.

Signed-off-by: Marcus Watts <mwatts@redhat.com>