openssl - jumbo patch.
// Do not commit this to master - apply the 3 patches in PR #8339 instead.

rgw / openssl - port_str split fix.
For civetweb: don't complain about trailing 's' on port number.
Document that radosgw now supports SSL.

ceph PR #8339 contains these 3 fixes, as 3 separate patches.
This is the same thing except as one big patch, with this
slightly more confusing commit message.

split port string.
7b453bed "Replace + with , in civetweb port string."

Remove code that complains if port is not a simple integer.

Civetweb can listen on more than one port, and port numbers can
be suffixed with 's' to indicate ssl.  There is already code elsewhere
to fetch the port string and pass it directly to civetweb.  So
the code here that parsed the port string and complained if it
wasn't a simple integer has not been doing anything useful for
a while.

This includes information on file format and configuration file syntax.

Signed-off-by: Marcus Watts <mwatts@redhat.com>