Use "mg_get_local_addr" to correctly deduce host port.
civetweb can accept connections on multiple ports, some of
which might have SSL turned on and some not.  Both s3 and
swift have various authorization protocols in which the
port number matters.  In the generic radosgw frontend process,
each frontend only has one port number, but we should
want to have both ssl and non-ssl connections managed
within one rgw frontend, because the thread pool is also
per front-end, and that *is* a scarce resource.

So this patch adds logic to use the results of mg_get_local_addr()
in preference to the generic front-end port number.  This will affect
"v4" s3 authorization, and also affect swift pre-signed URLs.

Signed-off-by: Marcus Watts <mwatts@redhat.com>