Store swift acls with keystone elements.
In swift, acls can contain keystone elements.  Keystone elements always
have project : user where project & user can be names (deprecated), ids,
or '*'.  In the master branch of ceph, these are stored as acl elements
with an empty name field and an id string that contains the element
string.  This is done with a bit of code that says "skipping silently".
(In jewel they really were being ignored silently.)

This code adds logic to emulate the same behavior in jewel.  This is
just the store logic for that, not the actual permissions logic.

The master branch also has logic to parse and store "referral" grants.
I've included commented out code to do the same.  It's commented out
because the jewel version of the grant logic doesn't have 'set_referer'
or type ACL_TYPE_REFERER, and I'm not sure I will ever need this logic
(I'm only including it because I don't want to pick this logic out of
master again.)

Signed-off-by: Marcus Watts <mwatts@redhat.com>