Enable 3rd party access to swift for most operations.
In order for cross-tenant swift acls to be useful, it's necessary to
enable "3rd party" access.  This means parsing the URL for the desired
account and using that account rather than the user's to locate the
desired bucket.  Existing code was already capable of parsing the account
out of the URL, but failed to use it when accessing the bucket.

This fix is jewel specific; master has a fix for this, but the relevant
code path is quite different.

Signed-off-by: Marcus Watts <mwatts@redhat.com>