Reject invalid swift acls.
EXPERIMENTAL
This version of the logical also no longer skips silently; it should
flag and return an error if it doesn't like an acl.
XXX This may change.  (* really ignore?  configuration option to change?? *)

Note that there was existing logic to check to see if the result
from RGWAccessControlPolicy_SWIFT::create was less than 0; this
could never previously happen because "boolean" true is 1, not -1.
This change makes the existing error case logic functional.

Signed-off-by: Marcus Watts <mwatts@redhat.com>