Changes to build ceph with a newer version of civetweb/openssl.
Civetweb 1.11 includes openssl 1.1.0 support, which is desirable
on recent versions of fedora.  Another desirable feature of
openssl 1.1.0 is that it omits the locking code which is otherwise
a pesky integration problem with curl.

To accomodate a newer version of ceph, it's necessary to change
how "info->http_headers" is referenced.  In older versions of
civetweb this was a nested structure which manifests in c++
as "mg_request_info::mg_header".  In civetweb 1.11, this
structure is no longer nested and its proper type is "mg_header".

When building civetweb.c, it is necessary to pass in '-DOPENSSL_API_1_1",
but only when actually linking against openssl 1.1.0.  With older
versions of openssl, this flag should not be passed in.

This set of commits does not require that the newer version
of civetweb or openssl be present.  It will also allow the use of
the newer civetweb with an older openssl, or an older version of
civetweb (only against an older openssl).

Signed-off-by: Marcus Watts <mwatts@redhat.com>