Fix tag handling for ceph_versioning.
The build system for ceph wants to store the revision and
git description of the ceph source.  The annotated tag that
this worked against were of the form "v10.2.7".  So the
logic used "cut -c 2-" to discard the v.  Later logic looks
at the tag in ceph_ver.h and constructs sed commands using the
delimiter / to insert the git version into what will become the
python executable "ceph".

I have to work with a ceph source tree that contains tags
that do not start with v and contain embedded tags.  An
example tag of this sort is "mdw/jewel-10.2.10".  The cut
above would discard the "m" in mdw.  The later logic would
construct invalid sed commands because of the /, which would
cause the build to fail.

So the changes here are to use sed to discard a leading v,
and to revise the later sed logic to more efficiently construct
a sed expression to insert git versioning.

Signed-off-by: Marcus Watts <mwatts@redhat.com>