rgw: bucket link: Add ability to name bucket w/ different tenant.
This is not a complete fix; but it makes it possible for the
bucket link command to correctly find and attempt to link a bucket
to a user with a different tenant.  The reason this is not a complete
fix is that with just this change, the resulting bucket is "broken";
a duplicate endpoint but 0 length contents is created, and
the info entry is not correctly moved.

Signed-off-by: Marcus Watts <mwatts@redhat.com>