rgw: sts: fixups: add LocalApplier ancestor back to STSEngine, fixup LocalApplier parms.
This is a collection of fixes to make the STS code compile
with the version of master I'm using.

One special change not present elsewhere:
LocalApplier has some extra optional arguments (policy roles, perm mask for token)
which are added for STSEngine.  Here I've added default arguments for
these arguments so that some existing code need not change from master.

Signed-off-by: Marcus Watts <mwatts@redhat.com>