rgw: STS recv_body fixes.
It's naughty to read data sized by the client into a small
fixed size buffer.  It's even more naughty to pretend it's
a null terminated C string and emit it in the log.  The result
of this is to emit random amounts of left over stack crud as
binary glop at the end of various strings.  The random binary
glop probably also winds up being appended to one of parameters,
which seems liable to cause future problems.

This fix: expands any binary crud in strings in the log file
as printable ascii sequences.  Also it fixes the body receive
so it won't overflow the stack and will properly nul terminate
the resulting string (and any resulting parameter values.)

// safe to collapse this commit into one of the STS commits; this
//	commit message should be elided when that's done.

Signed-off-by: Marcus Watts <mwatts@redhat.com>