rgw: Re-enable partial fetch of head object when GET request with range header
If a client specifies a byte range, it is still desirable to
prefetch from the head object, but it is not necessary to fetch an
entire chunk's worth of data.

This commit also includes partial support for multiple byte ranges.
The standard allows this.  This implementation does not attempt
to deliver multiple output parts, so if there are mutiple byte
ranges, a single part containing the entire byte range will be
returned.

Fixes: https://tracker.ceph.com/issues/44508
Signed-off-by: Marcus Watts <mwatts@redhat.com>