radosgw / AWSv4ComplMulti: better parsing for chunk header fields.
The existing logic for chunk headers assumes there will be ony one field
specified for chunk header, and that that field will be named will be
"chunk-signature".  RFC 2616/7230 make it clear it's possible to put
other fields there, so there may be S3 clients now or in the future that
do so.  Additionally Amazon could in the future extend S3 to make use of
this feature of the protocol.

This patch adds more robust parsing of the fields in the chunk header.
It should be capable of handling a more diverse mix of present and
future clients.

Fixes: https://tracker.ceph.com/issues/45790

Signed-off-by: Marcus Watts <mwatts@redhat.com>