rgw/beast: must close after body read fails.
With transfer-encoding: chunked, it is possible for bad data
to cause a the body read to fail early.  If this happens, the
input pointer will be left pointing to the middle of the data
that could not be consumed, which will cause subsequent attempts
to read a request on this connection to see the wrong data.
This is not useful, and it can lead to worse problems.

The only safe recovery choice is to close the connection
after issuing an error message.  This commit adds support to do this.

Fixes: https://tracker.ceph.com/issues/45789

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 8cb184d3089fb80f699189abb3b64900f791d7f0)

Conflicts:
	src/rgw/rgw_asio_frontend.cc
- StreamIO constructor lacked yield argument.