rgw/beast: handle boto2 chunk headers.
boto2 supports transfer-encoding: chunked, but when it
does so, it emits chunk headers that consist of the
hexadecimal count followed by a single ';'.  This violates
the documented standard which says the semi-colon should
either be omitted or a fieldname should be supplied
after that.  Older versions of ceph accepted this bad
header.  Recent versions of civetweb broke this, but I
have another fix to make this work with civetweb, and
this is the companion fix to make it work for beast as well.

This commit must be linked against a special version of beast
with the underlying support for this feature.

Fixes: https://tracker.ceph.com/issues/45789

Signed-off-by: Marcus Watts <mwatts@redhat.com>