First part copy object encryption fixes.
This is most of the first phase of the copy object encryption fixes.
It comprises additional data paths to communicate data from the
rest layer down to the sal layer to handle decrypting
objects.

XXX This version compiles, but there are a few loose ends
XXX to tidy yet so this is not quite funtional.

Past this, a future set of commits based on this needs to use the same
data path to pass down re-encrypt data for the destination.

Signed-off-by: Marcus Watts <mwatts@redhat.com>