rgw: resolve some link problems between rgw_a and rgw_common
This is after a suggestion by Matt Benjamin.

Not sure what caused this problem, but in this branch,
rgw_op.cc.o, which is in rgw_common.a, requires these symbols:
	ExternalTokenEngine::authenticate
	keystone::TokenEngine::authenticate
	rgw::auth::swift::SignedTokenEngine::authenticate
	vtable for EC2Engine
	vtable for TempURLApplier
	vtable for TempURLEngine
these are defined in the object files for these in rgw_a:
	rgw_swift_auth.cc
	rgw_auth_keystone.cc
Additionally, the swift changes also require,
	rgw_rest_swift.cc
So this change mvoes those files from rgw_a to rgw_common.

Signed-off-by: Marcus Watts <mwatts@redhat.com>